move load to operator
